test(int): add integrate test suites for final gen
